+++ CoreMIDI4J/src/uk/co/xfactorylibrarians/coremidi4j/CoreMidiDeviceProvider.java
@@ -1,4 +1,4 @@
-/**
+/*
  * Title:        CoreMIDI4J
  * Description:  Core MIDI Device Provider for Java on OS X
  * Copyright:    Copyright (c) 2015-2016
@@ -32,7 +32,7 @@ public class CoreMidiDeviceProvider extends MidiDeviceProvider implements CoreMi
 
     private CoreMidiClient client;
     private CoreMidiOutputPort output;
-    private Map<Integer, MidiDevice> deviceMap = new LinkedHashMap<Integer, MidiDevice>(DEVICE_MAP_SIZE);
+    private final Map<Integer, MidiDevice> deviceMap = new LinkedHashMap<>(DEVICE_MAP_SIZE);
 
   }
 
@@ -41,7 +41,7 @@ public class CoreMidiDeviceProvider extends MidiDeviceProvider implements CoreMi
   /**
    * Initialises the system
    * 
-   * @throws CoreMidiException Thrown if error occurs.
+   * @throws CoreMidiException if there is a problem communicating with CoreMIDI
    * 
    */
 
@@ -82,13 +82,13 @@ public class CoreMidiDeviceProvider extends MidiDeviceProvider implements CoreMi
   /**
    * Builds the device map
    * 
-   * @throws CoreMidiException Thrown if error occurs
+   * @throws CoreMidiException if there is a problem communicating with CoreMIDI
    * 
    */
 
   private void buildDeviceMap() throws CoreMidiException {
 
-    Set<Integer> devicesSeen = new HashSet<Integer>();
+    Set<Integer> devicesSeen = new HashSet<>();
 
     // Iterate through the sources
     for (int i = 0; i < getNumberOfSources(); i++) {
@@ -128,14 +128,34 @@ public class CoreMidiDeviceProvider extends MidiDeviceProvider implements CoreMi
 
     }
 
-    // Finally, remove any devices from the map which were no longer available according to CoreMIDI.
-    Set<Integer> devicesInMap = new HashSet<Integer>(midiProperties.deviceMap.keySet());
+    // Finally, remove any devices from the map which were no longer available according to CoreMIDI, and close them
+    // appropriately as needed.
+    Set<Integer> devicesInMap = new HashSet<>(midiProperties.deviceMap.keySet());
 
     for (Integer uniqueID : devicesInMap) {
 
       if ( !devicesSeen.contains(uniqueID) ) {
 
-        midiProperties.deviceMap.remove(uniqueID);
+        MidiDevice vanishedDevice = midiProperties.deviceMap.remove(uniqueID);
+
+        try {
+
+          if (vanishedDevice instanceof CoreMidiSource) {
+
+            // Must handle specially to avoid trying to interact with defunct CoreMIDI device
+            ((CoreMidiSource) vanishedDevice).deviceDisappeared();
+
+          } else {
+
+            vanishedDevice.close();  // CoreMidiDestination close is safe to call even after the device is gone
+
+          }
+
+        } catch (Exception e) {
+
+          System.err.println("Problem trying to clean up vanished MIDI device " + vanishedDevice + ": " + e);
+          e.printStackTrace();
+        }
 
       }
 
@@ -148,7 +168,7 @@ public class CoreMidiDeviceProvider extends MidiDeviceProvider implements CoreMi
    * 
    * @return	The CoreMidiClient object 
    * 
-   * @throws 	CoreMidiException Thrown if error occurs
+   * @throws 	CoreMidiException if there is a problem communicating with CoreMIDI
    * 
    */
 
@@ -223,7 +243,7 @@ public class CoreMidiDeviceProvider extends MidiDeviceProvider implements CoreMi
 
       final MidiDevice device = iterator.next();
 
-      info[counter] = (CoreMidiDeviceInfo) device.getDeviceInfo();
+      info[counter] = device.getDeviceInfo();
 
       counter += 1;
 
@@ -255,7 +275,7 @@ public class CoreMidiDeviceProvider extends MidiDeviceProvider implements CoreMi
 
     }
 
-    return (MidiDevice) midiProperties.deviceMap.get(((CoreMidiDeviceInfo) info).getUniqueID());
+    return midiProperties.deviceMap.get(((CoreMidiDeviceInfo) info).getUniqueID());
 
   }
 
@@ -414,7 +434,7 @@ public class CoreMidiDeviceProvider extends MidiDeviceProvider implements CoreMi
 
       if (isLibraryLoaded()) {
 
-        List<MidiDevice.Info> workingDevices = new ArrayList<MidiDevice.Info>(allInfo.length);
+        List<MidiDevice.Info> workingDevices = new ArrayList<>(allInfo.length);
         for (MidiDevice.Info candidate : allInfo) {
 
           try {
@@ -453,7 +473,7 @@ public class CoreMidiDeviceProvider extends MidiDeviceProvider implements CoreMi
   ///// JNI Interfaces
   //////////////////////////////
 
-  /**
+  /*
    * Static initializer for loading the native library
    *
    */
@@ -523,7 +543,7 @@ public class CoreMidiDeviceProvider extends MidiDeviceProvider implements CoreMi
    * 
    * @return    The UID of the referenced object
    * 
-   * @throws     CoreMidiException  Thrown if MIDI error occurs.
+   * @throws 						CoreMidiException if there is a problem communicating with CoreMIDI
    * 
    */