+++ CoreMIDI4J/src/uk/co/xfactorylibrarians/coremidi4j/CoreMidiSource.java
@@ -1,4 +1,4 @@
-/**
+/*
  * Title:        CoreMIDI4J
  * Description:  Core MIDI Device Provider for Java on OS X
  * Copyright:    Copyright (c) 2015-2016
@@ -14,10 +14,10 @@
 
 package uk.co.xfactorylibrarians.coremidi4j;
 
-import java.util.ArrayList;
-import java.util.Iterator;
-import java.util.List;
-import java.util.Vector;
+import java.util.*;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.concurrent.atomic.AtomicReference;
 
 import javax.sound.midi.InvalidMidiDataException;
 import javax.sound.midi.MidiDevice;
@@ -36,9 +36,9 @@ import javax.sound.midi.Transmitter;
 public class CoreMidiSource implements MidiDevice {
 
   private final CoreMidiDeviceInfo info;
-  private boolean isOpen;
-  private CoreMidiInputPort input = null;
-  private final List<Transmitter> transmitters;
+  private final AtomicBoolean isOpen;
+  private final AtomicReference<CoreMidiInputPort> input;
+  private final Set<CoreMidiTransmitter> transmitters;
 
   private int currentMessage = 0;  								// Will contain the status byte (> 127) while gathering a multi-byte message.
   private boolean currentDataIsSingleByte;  			// Is true if currentMessage only needs one byte of data.
@@ -53,15 +53,15 @@ public class CoreMidiSource implements MidiDevice {
    *
    * @param info a CoreMidiDeviceInfo object providing details of the MIDI interface
    * 
-   * @throws CoreMidiException Thrown if MIDI error occurs.
    * 
    */
 
-  CoreMidiSource(CoreMidiDeviceInfo info) throws CoreMidiException {
+  CoreMidiSource(CoreMidiDeviceInfo info) {
 
     this.info = info;
-    this.isOpen = false;
-    transmitters = new ArrayList<Transmitter>();
+    input = new AtomicReference<>();
+    isOpen = new AtomicBoolean(false);
+    transmitters = Collections.newSetFromMap(new ConcurrentHashMap<CoreMidiTransmitter, Boolean>());
 
   }
 
@@ -89,65 +89,88 @@ public class CoreMidiSource implements MidiDevice {
   @Override
   public void open() throws MidiUnavailableException {
 
-    try {
+    if (isOpen.compareAndSet(false, true)) {
 
-      // Create the input port if not already created
-      if (this.input == null) {
+      try {
 
-        this.input = CoreMidiDeviceProvider.getMIDIClient().inputPortCreate("Core Midi Provider Input");
+        // Create the input port if not already created
+        if (input.get() == null) {
+
+          input.set(CoreMidiDeviceProvider.getMIDIClient().inputPortCreate("Core Midi Provider Input"));
+
+        }
+
+        // And connect to it
+        input.get().connectSource(this);
+
+        // Get the system time in microseconds
+        startTime = this.getMicroSecondTime();
+
+      } catch (CoreMidiException e) {
+
+        e.printStackTrace();
+        throw new MidiUnavailableException(e.getMessage());
 
       }
-
-      // And connect to it
-      this.input.connectSource(this);
-      isOpen = true;
-      
-      // Get the system time in microseconds
-      startTime = this.getMicroSecondTime();
-
-    } catch (CoreMidiException e) {
-
-      e.printStackTrace();
-      throw new MidiUnavailableException(e.getMessage());
 
     }
 
   }
 
   /**
-   * Closes the Core MIDI Device
+   * Closes the Core MIDI Device, which also closes all its transmitters
    * 
    */
 
   @Override
   public void close() {
 
-    try {
+    if (isOpen.compareAndSet(true, false)) {
 
-      // If the port is created then disconnect from it
-      if (this.input != null) {
+      try {
 
-        this.input.disconnectSource(this);
+        // If the port is created then disconnect from it
+        if (input.get() != null) {
+
+          try {
+
+            input.get().disconnectSource(this);
+
+          } finally {
+
+            input.set(null);
+
+          }
+
+        }
+
+        // Close all our transmitters, which will also clear the list.
+        // We iterate on a copy of the transmitter list to avoid issues with concurrent modification.
+        for (Transmitter transmitter : getTransmitters()) {
+
+          transmitter.close();
+
+        }
+
+      } catch (CoreMidiException e) {
+
+        e.printStackTrace();
 
       }
-
-      // Clear the transmitter list
-      synchronized (transmitters) {
-
-        transmitters.clear();
-
-      }
-
-    } catch (CoreMidiException e) {
-
-      e.printStackTrace();
-
-    } finally {
-
-      // Reset the context data
-      isOpen = false;
 
     }
+
+  }
+
+  /**
+   * Forcibly close because the underlying CoreMIDI device has disappeared. Behaves like {@link #close()} without
+   * attempting to detach from the now-nonexistent underlying device.
+   */
+
+  void deviceDisappeared() {
+
+    input.set(null);
+    close();
 
   }
 
@@ -163,7 +186,7 @@ public class CoreMidiSource implements MidiDevice {
   @Override
   public boolean isOpen() {
 
-    return isOpen;
+    return isOpen.get();
 
   }
 
@@ -237,7 +260,7 @@ public class CoreMidiSource implements MidiDevice {
   /**
    * Gets a list of receivers connected to the device
    *
-   * @return NULL - we do not maintain a list of receivers
+   * @return an empty list - we do not maintain a list of receivers
    * 
    * @see javax.sound.midi.MidiDevice#getReceivers()
    * 
@@ -247,12 +270,12 @@ public class CoreMidiSource implements MidiDevice {
   public List<Receiver> getReceivers() {
 
     // A CoreMidiSource has no receivers
-    return null;
+    return Collections.emptyList();
 
   }
 
   /**
-   * Gets a transmitter for this device (which is also added to the internal list
+   * Gets a transmitter for this device (which is also added to the internal list)
    *
    * @return a transmitter for this device
    * 
@@ -267,14 +290,10 @@ public class CoreMidiSource implements MidiDevice {
   public Transmitter getTransmitter() throws MidiUnavailableException {
 
     // Create the transmitter
-    Transmitter transmitter = new CoreMidiTransmitter(this);
+    CoreMidiTransmitter transmitter = new CoreMidiTransmitter(this);
 
-    // Add it to the list
-    synchronized (transmitters) {
-
-      transmitters.add(transmitter);
-
-    }
+    // Add it to the set of open transmitters
+    transmitters.add(transmitter);
 
     // Finally return it
     return transmitter;
@@ -282,9 +301,21 @@ public class CoreMidiSource implements MidiDevice {
   }
 
   /**
+   * Reacts to the closing of a transmitter by removing it from the set of active transmitters
+   *
+   * @param transmitter the transmitter which is reporting itself as having closed
+   */
+
+  void transmitterClosed(CoreMidiTransmitter transmitter) {
+
+    transmitters.remove(transmitter);
+
+  }
+
+  /**
    * Gets the list of transmitters registered with this MIDI device
    *
-   * @return The list of transmitters registered with this MIDI device
+   * @return The list of transmitters created from this MIDI device that are still open
    * 
    * @see javax.sound.midi.MidiDevice#getTransmitters()
    * 
@@ -293,16 +324,8 @@ public class CoreMidiSource implements MidiDevice {
   @Override
   public List<Transmitter> getTransmitters() {
 
-    // Create and return a list of transmitters
-    synchronized (transmitters) {
-
-      final List<Transmitter> list = new ArrayList<Transmitter>();
-
-      list.addAll(transmitters);
-
-      return list;
-
-    }
+    // Return an immutable copy of our current set of open transmitters
+    return Collections.unmodifiableList(new ArrayList<Transmitter>(transmitters));
 
   }
 
@@ -315,6 +338,7 @@ public class CoreMidiSource implements MidiDevice {
    * @return true if the status byte represents a standalone real-time message which should be passed on without
    *         interrupting any other message being gathered.
    */
+
   private boolean isRealTimeMessage(byte status) {
 
     switch ( status ) {
@@ -511,7 +535,7 @@ public class CoreMidiSource implements MidiDevice {
           // We are starting to gather a SYSEX message.
           currentMessage = SysexMessage.SYSTEM_EXCLUSIVE;
           sysexMessageLength = 0;
-          sysexMessageData = new Vector<byte[]>();
+          sysexMessageData = new Vector<>();
           offset += processSysexData(packetlength, data, offset, timestamp);
 
         } else {
@@ -555,7 +579,7 @@ public class CoreMidiSource implements MidiDevice {
    *
    * @return The constructed SYSEX message
    * 
-   * @throws InvalidMidiDataException  Thrown if MIDI data malformed.
+   * @throws InvalidMidiDataException if the data is not properly formed
    * 
    */
 
@@ -566,11 +590,9 @@ public class CoreMidiSource implements MidiDevice {
     int index = 0;
 
     // Iterate through the partial messages
-    for (int i = 0; i < sysexMessageData.size(); i += 1) {
+    for (byte[] sourceData : sysexMessageData) {
 
       // Get the partial message
-      byte sourceData[] = sysexMessageData.get(i);
-
       // Copy the partial message into the array
       System.arraycopy(sourceData, 0, data, index, sourceData.length);
 
@@ -603,9 +625,9 @@ public class CoreMidiSource implements MidiDevice {
    * @param startOffset  	The position within the packet where the current message began
    * @param timestamp 		The message timestamp
    *
-   * @return 							The number of bytes consumed from the packet by the SYSEX message.
+   * @return 							The number of bytes consumed from the packet by the SYSEX message
    * 
-   * @throws     InvalidMidiDataException   Thrown if MIDI data malformed.
+   * @throws 							InvalidMidiDataException if the data is not properly formed
    * 
    */
 
@@ -710,30 +732,15 @@ public class CoreMidiSource implements MidiDevice {
     //
     //		}
 
-    synchronized (transmitters) {
+    // Iterate over a snapshot of our transmitters to avoid issues with concurrent modification
+    for (Transmitter transmitter : getTransmitters()) {
 
-      // Get the iterators from the transmitters collection
-      final Iterator<Transmitter> iterator = transmitters.iterator();
+      final Receiver receiver = transmitter.getReceiver();
 
-      // Loop through the transmitters
-      while (iterator.hasNext()) {
+      // If the transmitter has a non-null receiver, then send the message
+      if (receiver != null) {
 
-        // Get the next transmitter
-        final Transmitter transmitter = iterator.next();
-
-        // If the transmitter is not null then get the receiver
-        if (transmitter != null) {
-
-          final Receiver receiver = transmitter.getReceiver();
-
-          // If the receiver is not null then get send the message
-          if (receiver != null) {
-
-            receiver.send(message, timestamp);
-
-          }
-
-        }
+        receiver.send(message, timestamp);
 
       }
 
@@ -744,22 +751,23 @@ public class CoreMidiSource implements MidiDevice {
   /**
    * Formats the provided data into a HEX string, which is useful for debugging
    *
-   * @param aByte The data to format
+   * @param aByteArray The data to format
    * 
    * @return The formatted HEX string
    * 
    */
 
-  private String getHexString(byte[] aByte) {
+  @SuppressWarnings("unused")
+  private String getHexString(byte[] aByteArray) {
 
-    StringBuffer sbuf = new StringBuffer(aByte.length * 3 + 2);
+    StringBuffer sbuf = new StringBuffer(aByteArray.length * 3 + 2);
 
-    for (int i = 0; i < aByte.length; i++) {
+    for (byte aByte : aByteArray) {
 
       sbuf.append(' ');
-      byte bhigh = (byte) ((aByte[i] & 0xf0) >> 4);
+      byte bhigh = (byte) ((aByte & 0xf0) >> 4);
       sbuf.append((char) (bhigh > 9 ? bhigh + 'A' - 10 : bhigh + '0'));
-      byte blow = (byte) (aByte[i] & 0x0f);
+      byte blow = (byte) (aByte & 0x0f);
       sbuf.append((char) (blow > 9 ? blow + 'A' - 10 : blow + '0'));
 
     }
@@ -772,7 +780,7 @@ public class CoreMidiSource implements MidiDevice {
   ///// JNI Interfaces
   //////////////////////////////
 
-  /**
+  /*
    * Static initializer for loading the native library
    *
    */