+++ CHANGELOG.md
@@ -6,7 +6,16 @@ This change log follows the conventions of
 
 ## [Unreleased][unreleased]
 
-Nothing so far.
+### Fixed
+
+- Calling `close()` more than once on a `CoreMidiSource` would lead to
+  a [crash](https://github.com/DerekCook/CoreMidi4J/issues/19) as the
+  same block of memory is attempted to be freed more than once.
+- Implemented the rest of the MIDI SPI contract for tracking and
+  closing transmitters and receivers.
+- We now close any devices (and their transmitters or receivers) when
+  the underlying CoreMIDI device disappears.
+
 
 ## [1.0] - 2017-05-14